requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.1
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
daphne==4.1.2
darkdetect==0.8.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.17.0
fonttools==4.56.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
randomcolor==0.4.4.6
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tkcalendar==1.6.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
websockets==15.0.1
zope.interface==7.2
